external-scaler/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
grpcio==1.62.1
grpcio-tools==1.62.1
idna==3.6
numpy==1.26.4
pandas==2.2.1
protobuf==4.25.3
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
